load tests in separate file & using sockets
